Remove existing Root endpoint and replace with extended
